Backend/Services/Input_Mgr/requirements.txt
flask
redis
kafka-python
